+++ react-app/src/main.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import "../../styles.css";
 import App from "./App";
+import "./legacy-root.css";
 import "./styles.css";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(